# php-tutor/php-learn/app/cache/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "php-learn API",
        "description": "PHP 8.5 无框架后端 + Vue 3 SPA 前后端分离项目的 REST API。\n\n鉴权方式：Bearer Token（JWT Access + Refresh Token）\n响应结构：统一 envelope {code, message, data, errors?}\n业务 code：0=成功, 40101=未登录, 40102=token 无效, 40301=无权限, 40401=资源不存在, 42201=校验失败, 50000=内部错误",
        "contact": {
            "name": "php-learn",
            "url": "https:\/\/github.com\/example\/php-learn"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "\/api\/v1",
            "description": "API v1（同域部署）"
        },
        {
            "url": "http:\/\/127.0.0.1:8000\/api\/v1",
            "description": "本地开发"
        }
    ],
    "paths": {
...
        "\/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "文章列表",
                "description": "支持分页 + 分类\/状态筛选",
                "operationId": "66bf6230eeef4cfa0a5c8e7a25e9715e",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "页码",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "每页条数",
                        "schema": {
                            "type": "integer",
                            "default": 8,
                            "maximum": 100
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "分类筛选（php\/java\/db\/other）",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "php",
                                "java",
                                "db",
                                "other"
                            ]
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "状态筛选（published\/draft）",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "draft",
                                "published"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "分页列表"
                    }
                }
            },
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "新建文章",
                "description": "需要 Bearer Token",
                "operationId": "e5e5b148ab57bfb1c951ddf23fec45d7",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "创建成功",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/Article"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "参数校验失败",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "\/articles\/{id}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "文章详情",
                "description": "获取指定 ID 的文章详情",
                "operationId": "b52c3be0bc4fedb09a4b12b9e4802165",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "文章 ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "文章详情",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/Article"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "文章不存在",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Articles"
                ],
                "summary": "更新文章",
                "description": "PATCH 语义：只更新提交了字段",
                "operationId": "822c46f61b39bce0ca3e8158393301f6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "文章 ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/Article"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "文章不存在",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "参数校验失败",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Articles"
                ],
                "summary": "删除文章",
                "description": "DELETE \/api\/v1\/articles\/{id}（需鉴权）",
                "operationId": "8ec6712f08b29739e503b70f0487584d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "文章 ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "删除成功（无 body）"
                    },
                    "401": {
                        "description": "未登录",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "文章不存在",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer Token（JWT Access Token），通过登录接口获取",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}